# operator-api/openapi.yaml
openapi: 3.1.0
info:
  title: Kernel Images API
  version: 0.1.0
paths:
...
  /fs/read_file:
    get:
      summary: Read file contents
      operationId: readFile
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            pattern: "^/.*"
          description: Absolute file path to read.
      responses:
        "200":
          description: File read successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/write_file:
    put:
      summary: Write or create a file
      operationId: writeFile
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            pattern: "^/.*"
          description: Destination absolute file path.
        - name: mode
          in: query
          required: false
          schema:
            type: string
            pattern: "^[0-7]{3,4}$"
          description: Optional file mode (octal string, e.g. 644). Defaults to 644.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: File written successfully
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/list_files:
    get:
      summary: List files in a directory
      operationId: listFiles
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            pattern: "^/.*"
          description: Absolute directory path.
      responses:
        "200":
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListFiles"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/create_directory:
    put:
      summary: Create a new directory
      operationId: createDirectory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDirectoryRequest"
      responses:
        "201":
          description: Directory created successfully
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/delete_file:
    put:
      summary: Delete a file
      operationId: deleteFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeletePathRequest"
      responses:
        "200":
          description: File deleted
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/delete_directory:
    put:
      summary: Delete a directory
      operationId: deleteDirectory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeletePathRequest"
      responses:
        "200":
          description: Directory deleted
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/set_file_permissions:
    put:
      summary: Set file or directory permissions/ownership
      operationId: setFilePermissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetFilePermissionsRequest"
      responses:
        "200":
          description: Permissions updated
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/file_info:
    get:
      summary: Get information about a file or directory
      operationId: fileInfo
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            pattern: "^/.*"
          description: Absolute path of the file or directory.
      responses:
        "200":
          description: File information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileInfo"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/move:
    put:
      summary: Move or rename a file or directory
      operationId: movePath
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovePathRequest"
      responses:
        "200":
          description: Move successful
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/watch:
    post:
      summary: Watch a directory for changes
      operationId: startFsWatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartFsWatchRequest"
      responses:
        "201":
          description: Watch started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  watch_id:
                    type: string
                    description: Unique identifier for the directory watch
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/watch/{watch_id}/events:
    get:
      summary: Stream filesystem events for a watch
      operationId: streamFsEvents
      parameters:
        - name: watch_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE stream of filesystem events
          headers:
            X-SSE-Content-Type:
              description: Media type of SSE data events (application/json)
              schema:
                type: string
                const: application/json
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/FileSystemEvent"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /fs/watch/{watch_id}:
    delete:
      summary: Stop watching a directory
      operationId: stopFsWatch
      parameters:
        - name: watch_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Watch stopped successfully
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"

  # --------------------------------
  # COMPAT ALIASES FOR XDOTool START
  # --------------------------------
...
  /fs/upload:
    post:
      summary: Upload a file into the VM
      operationId: fsUpload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [path, file]
              properties:
                path: { type: string, description: Destination absolute path }
                file: { type: string, format: binary }
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /fs/download:
    get:
      summary: Download a file from the VM
      operationId: fsDownload
      parameters:
        - in: query
          name: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: File bytes
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/NotFoundError"
  /fs/tail/stream:
    get:
      summary: Tail-follow a file as SSE
      operationId: fsTailStream
      parameters:
        - in: query
          name: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: text/event-stream of lines
          headers:
            X-SSE-Content-Type:
              description: Media type of SSE data events (application/json)
              schema:
                type: string
                const: application/json
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: data
                  data: {"line":"hello","ts":"2025-08-09T10:00:00Z"}
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /bus/publish:
    post:
      summary: Publish a message to in-VM agent subscribers
      operationId: busPublish
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BusPublishRequest" }
      responses:
        "200":
          description: Delivered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BusPublishResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /bus/subscribe:
    get:
      summary: Subscribe to agent-emitted messages (SSE)
      operationId: busSubscribe
      parameters:
        - in: query
          name: channel
          required: true
          schema: { type: string, example: "agent" }
      responses:
        "200":
          description: text/event-stream of messages
          headers:
            X-SSE-Content-Type:
              description: Media type of SSE data events (application/json)
              schema:
                type: string
                const: application/json
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: data
                  data: {"ts":"2025-08-09T10:00:00Z","type":"custom_event","payload":{"ok":true}}
...
  /browser/har/start:
    post:
      summary: Start capturing HAR from default browser profile
      operationId: browserHarStart
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BrowserHarStartRequest" }
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrowserHarStartResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /browser/har/{har_session_id}/stream:
    get:
      summary: Stream HAR entries for a HAR session (SSE)
      operationId: browserHarStream
      parameters:
        - in: path
          name: har_session_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: text/event-stream of HAR entries
          headers:
            X-SSE-Content-Type:
              description: Media type of SSE data events (application/json)
              schema:
                type: string
                const: application/json
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: data
                  data: {"ts":"2025-08-09T10:00:00Z","entry":{"request":{"method":"GET","url":"https://..."}}}
        "404":
          $ref: "#/components/responses/NotFoundError"
  /browser/har/stop:
    post:
      summary: Stop HAR capture
      operationId: browserHarStop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [har_session_id]
              properties:
                har_session_id: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    StartFsWatchRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
          description: Directory to watch.
        recursive:
          type: boolean
          description: Whether to watch recursively.
          default: false
      additionalProperties: false
    FileSystemEvent:
      type: object
      description: Filesystem change event.
      required: [type, path]
      properties:
        type:
          type: string
          enum: [CREATE, WRITE, DELETE, RENAME]
          description: Event type.
        name:
          type: string
          description: Base name of the file or directory affected.
        path:
          type: string
          description: Absolute path of the file or directory.
        is_dir:
          type: boolean
          description: Whether the affected path is a directory.
...
    FileInfo:
      type: object
      required: [name, path, size_bytes, is_dir, mod_time, mode]
      properties:
        name:
          type: string
          description: Base name of the file or directory.
        path:
          type: string
          description: Absolute path.
        size_bytes:
          type: integer
          description: Size in bytes. 0 for directories.
        is_dir:
          type: boolean
          description: Whether the path is a directory.
        mod_time:
          type: string
          format: date-time
          description: Last modification time.
        mode:
          type: string
          description: File mode bits (e.g., "drwxr-xr-x" or "-rw-r--r--").
    ListFiles:
      type: array
      description: Array of file or directory information entries.
      items:
        $ref: "#/components/schemas/FileInfo"
    CreateDirectoryRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
          description: Absolute directory path to create.
          pattern: "^/.*"
        mode:
          type: string
          description: Optional directory mode (octal string, e.g. 755). Defaults to 755.
          pattern: "^[0-7]{3,4}$"
      additionalProperties: false
    DeletePathRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
          description: Absolute path to delete.
          pattern: "^/.*"
      additionalProperties: false
    SetFilePermissionsRequest:
      type: object
      required: [path, mode]
      properties:
        path:
          type: string
          description: Absolute path whose permissions are to be changed.
          pattern: "^/.*"
        owner:
          type: string
          description: New owner username or UID.
        group:
          type: string
          description: New group name or GID.
        mode:
          type: string
          description: File mode bits (octal string, e.g. 644).
          pattern: "^[0-7]{3,4}$"
      additionalProperties: false
    MovePathRequest:
      type: object
      required: [src_path, dest_path]
      properties:
        src_path:
          type: string
          description: Absolute source path.
          pattern: "^/.*"
        dest_path:
          type: string
          description: Absolute destination path.
          pattern: "^/.*"
      additionalProperties: false

    # -------------------------
    # NEW SCHEMAS (ADDITIVE)
    # -------------------------
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
...
    UploadResponse:
      type: object
      properties:
        path: { type: string }
        size_bytes: { type: integer }
...
    BusPublishRequest:
      type: object
      properties:
        channel: { type: string, example: agent }
        type: { type: string, example: custom_event }
        payload: { type: object }
      required: [channel, type, payload]
    BusPublishResponse:
      type: object
      properties:
        delivered: { type: boolean }
...
    BrowserHarStartRequest:
      type: object
      properties:
        include_content: { type: boolean, default: true }
    BrowserHarStartResponse:
      type: object
      properties:
        har_session_id: { type: string, format: uuid }
...
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"